fix: split ListUsbDevices into platform-specific files to silence macOS error
On macOS, the usbDeviceMonitor was logging an error every 5s:
  [storage] usbDeviceMonitor: failed to list USB devices: open /sys/bus/usb/devices/: no such file or directory

Split ListUsbDevices (and the Linux-only isDeviceMounted helper) into
build-tagged files following the existing detector_darwin.go / detector_linux.go
pattern:

- usb_devices_linux.go  (//go:build linux)  — original /sys/bus/usb/devices/ impl
- usb_devices_darwin.go (//go:build darwin) — returns ([]UsbDevice{}, nil) silently
- usb_devices_other.go  (//go:build !linux && !darwin) — same no-op stub

Closes #1280
